+++ Frontend/app/shared/animations/loading-screen.animation.ts
@@ -1,9 +1,8 @@
-// taken from https://codepen.io/IvanKhartov/pen/KmgzpX
-import {Component, Input, OnChanges, SimpleChanges} from '@angular/core';
+import {Component, Input, OnChanges, OnInit, SimpleChanges} from '@angular/core';
 @Component({
 	selector: `data-loading-screen`,
 	template: `
-		<div *ngIf="!remove" class="loading-wrapper" [class.data-ready]="dataReady">
+		<div *ngIf="!remove" class="loading-wrapper" [class.data-ready]="!!dataReady" >
 				<div class="diamonds">
 				<div class="diamond diamond-1"></div>
 				<div class="diamond diamond-2"></div>
@@ -14,16 +13,22 @@ import {Component, Input, OnChanges, SimpleChanges} from '@angular/core';
 	`,
 })
 
-export class DiamondLoaderComponent implements OnChanges {
+export class DiamondLoaderComponent implements OnChanges, OnInit {
 	@Input()
 	public dataReady: boolean = false;
 	public remove: boolean = false;
 
+	public ngOnInit(): void {
+	}
+
 	public ngOnChanges(simpleChanges: SimpleChanges): void {
-		if(simpleChanges.dataReady && simpleChanges.dataReady.currentValue === true){
-			setTimeout(()=>{
+		if (simpleChanges.dataReady && simpleChanges.dataReady.currentValue === true) {
+			console.log('hit onChange', simpleChanges);
+			setTimeout(() => {
 				this.remove = true;
-			}, 600)
+			}, 1000)
 		}
 	}
+
+
 }