+++ Frontend/app/shared/shared.module.ts
@@ -9,8 +9,10 @@ import {ModelService} from './services/model.service';
 import {LastIndexed} from './pipes/lastIndex.pipe';
 import {SelectByID} from './pipes/selectByID.pipe';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {ConfirmationComponent} from './components/confirmation.component';
 
 const COMPONENTS = [
+	ConfirmationComponent,
 	LastIndexed,
 	Capitalize,
 	SelectByID,
@@ -19,7 +21,6 @@ const COMPONENTS = [
 
 const MODULES = [
 	BrowserAnimationsModule,
-	// NoopAnimationsModule,
 	AngularCommonModule,
 	ReactiveFormsModule,
 	FormsModule,